Add summarization experiments (#173)
* Merge pull request #167 from microsoft/roblou/small-mole

Add summarization experiments for release

* Fix bad merge
