+++ src/extension/intents/node/agentIntent.ts
@@ -12,7 +12,6 @@ import { ChatLocation, ChatResponse } from '../../../platform/chat/common/common
 import { ConfigKey, IConfigurationService } from '../../../platform/configuration/common/configurationService';
 import { modelMightUseReplaceStringExclusively, modelSupportsApplyPatch, modelSupportsReplaceString } from '../../../platform/endpoint/common/chatModelCapabilities';
 import { IEndpointProvider } from '../../../platform/endpoint/common/endpointProvider';
-import { CacheType } from '../../../platform/endpoint/common/endpointTypes';
 import { IEnvService } from '../../../platform/env/common/envService';
 import { ILogService } from '../../../platform/log/common/logService';
 import { IEditLogService } from '../../../platform/multiFileEdit/common/editLogService';
@@ -44,6 +43,7 @@ import { EditCodePrompt2 } from '../../prompts/node/panel/editCodePrompt2';
 import { ToolResultMetadata } from '../../prompts/node/panel/toolCalling';
 import { ToolName } from '../../tools/common/toolNames';
 import { IToolsService } from '../../tools/common/toolsService';
+import { addCacheBreakpoints } from './cacheBreakpoints';
 import { EditCodeIntent, EditCodeIntentInvocation, EditCodeIntentInvocationOptions, mergeMetadata, toNewChatReferences } from './editCodeIntent';
 import { getRequestedToolCallIterationLimit, IContinueOnErrorConfirmation } from './toolCallingLoop';
 
@@ -294,85 +294,4 @@ export class AgentIntentInvocation extends EditCodeIntentInvocation {
 	}
 
 	override processResponse = undefined;
-}
-
-const MaxCacheBreakpoints = 4;
-
-/**
- * Prompt cache breakpoint strategy:
- *
- * The prompt is structured like
- * - System message
- * - Custom instructions
- * - Global context message (has prompt-tsx cache breakpoint)
- * - History
- * - Current user message with extra context
- * - Current tool call rounds
- *
- * Below the current user message, we add cache breakpoints to the last tool result in each round.
- * We add one to the current user message.
- * And above the current user message, we add breakpoionts to an assistant message with no tool calls (so the terminal response in a turn).
- *
- * There will always be a cache miss when a new turn starts because the previous messages move from below the current user message with extra context to above it.
- * For turns with no tool calling, we will have a hit on the previous assistant message in history.
- * During the agentic loop, each request will have a hit on the previous tool result message.
- */
-export function addCacheBreakpoints(messages: Raw.ChatMessage[]) {
-	// One or two cache breakpoints are already added via the prompt, assign the rest here.
-	let count = MaxCacheBreakpoints - countCacheBreakpoints(messages);
-	let isBelowCurrentUserMessage = true;
-	const reversedMsgs = [...messages].reverse();
-	for (const [idx, msg] of reversedMsgs.entries()) {
-		const prevMsg = reversedMsgs.at(idx - 1);
-		const hasCacheBreakpoint = msg.content.some(part => part.type === Raw.ChatCompletionContentPartKind.CacheBreakpoint);
-		if (hasCacheBreakpoint) {
-			continue;
-		}
-
-		const isLastToolResultInRound = msg.role === Raw.ChatRole.Tool && prevMsg?.role !== Raw.ChatRole.Tool;
-		const isAsstMsgWithNoTools = msg.role === Raw.ChatRole.Assistant && !msg.toolCalls?.length;
-		if (isBelowCurrentUserMessage && (isLastToolResultInRound || msg.role === Raw.ChatRole.User) || isAsstMsgWithNoTools) {
-			count--;
-			msg.content.push({
-				type: Raw.ChatCompletionContentPartKind.CacheBreakpoint,
-				cacheType: CacheType
-			});
-
-			if (count <= 0) {
-				break;
-			}
-		}
-
-		if (msg.role === Raw.ChatRole.User) {
-			isBelowCurrentUserMessage = false;
-		}
-	}
-
-	// If we still have cache breakpoints to allocate, add them from the system and custom instructions messages, if applicable.
-	for (const msg of messages) {
-		if (count <= 0) {
-			break;
-		}
-
-		const hasCacheBreakpoint = msg.content.some(part => part.type === Raw.ChatCompletionContentPartKind.CacheBreakpoint);
-		if ((msg.role === Raw.ChatRole.User || msg.role === Raw.ChatRole.System) && !hasCacheBreakpoint) {
-			count--;
-			msg.content.push({
-				type: Raw.ChatCompletionContentPartKind.CacheBreakpoint,
-				cacheType: CacheType
-			});
-		}
-
-		if (msg.role !== Raw.ChatRole.User && msg.role !== Raw.ChatRole.System) {
-			break;
-		}
-	}
-}
-
-function countCacheBreakpoints(messages: Raw.ChatMessage[]) {
-	let count = 0;
-	for (const msg of messages) {
-		count += msg.content.filter(part => part.type === Raw.ChatCompletionContentPartKind.CacheBreakpoint).length;
-	}
-	return count;
 }
