+++ src/extension/prompts/node/agent/summarizedConversationHistory.tsx
@@ -10,17 +10,22 @@ import { ChatMessage } from '@vscode/prompt-tsx/dist/base/output/rawTypes';
 import type { ChatResponsePart, LanguageModelToolInformation, NotebookDocument, Progress } from 'vscode';
 import { ChatFetchResponseType, ChatLocation, ChatResponse, FetchSuccess } from '../../../../platform/chat/common/commonTypes';
 import { ConfigKey, IConfigurationService } from '../../../../platform/configuration/common/configurationService';
+import { IEndpointProvider } from '../../../../platform/endpoint/common/endpointProvider';
 import { ILogService } from '../../../../platform/log/common/logService';
 import { IChatEndpoint } from '../../../../platform/networking/common/networking';
+import { APIUsage } from '../../../../platform/networking/common/openai';
 import { IPromptPathRepresentationService } from '../../../../platform/prompts/common/promptPathRepresentationService';
+import { IExperimentationService } from '../../../../platform/telemetry/common/nullExperimentationService';
 import { ITelemetryService } from '../../../../platform/telemetry/common/telemetry';
 import { IWorkspaceService } from '../../../../platform/workspace/common/workspaceService';
 import { CancellationToken } from '../../../../util/vs/base/common/cancellation';
 import { CancellationError, isCancellationError } from '../../../../util/vs/base/common/errors';
 import { Iterable } from '../../../../util/vs/base/common/iterator';
+import { StopWatch } from '../../../../util/vs/base/common/stopwatch';
 import { generateUuid } from '../../../../util/vs/base/common/uuid';
 import { IInstantiationService } from '../../../../util/vs/platform/instantiation/common/instantiation';
 import { ChatResponseProgressPart2 } from '../../../../vscodeTypes';
+import { addCacheBreakpoints } from '../../../intents/node/cacheBreakpoints';
 import { ToolCallingLoop } from '../../../intents/node/toolCallingLoop';
 import { IResultMetadata } from '../../../prompt/common/conversation';
 import { IBuildPromptContext, IToolCallRound } from '../../../prompt/common/intents';
@@ -30,14 +35,115 @@ import { NotebookSummary } from '../../../tools/node/notebookSummaryTool';
 import { renderPromptElement } from '../base/promptRenderer';
 import { Tag } from '../base/tag';
 import { ChatToolCalls } from '../panel/toolCalling';
-import { AgentUserMessage, getKeepGoingReminder, getUserMessagePropsFromAgentProps, getUserMessagePropsFromTurn } from './agentPrompt';
+import { AgentPrompt, AgentPromptProps, AgentUserMessage, getKeepGoingReminder, getUserMessagePropsFromAgentProps, getUserMessagePropsFromTurn } from './agentPrompt';
 import { SimpleSummarizedHistory } from './simpleSummarizedHistoryPrompt';
-import { StopWatch } from '../../../../util/vs/base/common/stopwatch';
-import { APIUsage } from '../../../../platform/networking/common/openai';
 
 export interface ConversationHistorySummarizationPromptProps extends SummarizedAgentHistoryProps {
 	simpleMode?: boolean;
 }
+
+const SummaryPrompt = <>
+	Your task is to create a comprehensive, detailed summary of the entire conversation that captures all essential information needed to seamlessly continue the work without any loss of context. This summary will be used to compact the conversation while preserving critical technical details, decisions, and progress.<br />
+
+	## Recent Context Analysis<br />
+
+	Pay special attention to the most recent agent commands and tool executions that led to this summarization being triggered. Include:<br />
+	- **Last Agent Commands**: What specific actions/tools were just executed<br />
+	- **Tool Results**: Key outcomes from recent tool calls (truncate if very long, but preserve essential information)<br />
+	- **Immediate State**: What was the system doing right before summarization<br />
+	- **Triggering Context**: What caused the token budget to be exceeded<br />
+
+	## Analysis Process<br />
+
+	Before providing your final summary, wrap your analysis in `&lt;analysis&gt;` tags to organize your thoughts systematically:<br />
+
+	1. **Chronological Review**: Go through the conversation chronologically, identifying key phases and transitions<br />
+	2. **Intent Mapping**: Extract all explicit and implicit user requests, goals, and expectations<br />
+	3. **Technical Inventory**: Catalog all technical concepts, tools, frameworks, and architectural decisions<br />
+	4. **Code Archaeology**: Document all files, functions, and code patterns that were discussed or modified<br />
+	5. **Progress Assessment**: Evaluate what has been completed vs. what remains pending<br />
+	6. **Context Validation**: Ensure all critical information for continuation is captured<br />
+	7. **Recent Commands Analysis**: Document the specific agent commands and tool results from the most recent operations<br />
+
+	## Summary Structure<br />
+
+	Your summary must include these sections in order, following the exact format below:<br />
+
+	<Tag name='analysis'>
+		[Chronological Review: Walk through conversation phases: initial request → exploration → implementation → debugging → current state]<br />
+		[Intent Mapping: List each explicit user request with message context]<br />
+		[Technical Inventory: Catalog all technologies, patterns, and decisions mentioned]<br />
+		[Code Archaeology: Document every file, function, and code change discussed]<br />
+		[Progress Assessment: What's done vs. pending with specific status]<br />
+		[Context Validation: Verify all continuation context is captured]<br />
+		[Recent Commands Analysis: Last agent commands executed, tool results (truncated if long), immediate pre-summarization state]<br />
+	</Tag><br />
+
+	<Tag name='summary'>
+		1. Conversation Overview:<br />
+		- Primary Objectives: [All explicit user requests and overarching goals with exact quotes]<br />
+		- Session Context: [High-level narrative of conversation flow and key phases]<br />
+		- User Intent Evolution: [How user's needs or direction changed throughout conversation]<br />
+
+		2. Technical Foundation:<br />
+		- [Core Technology 1]: [Version/details and purpose]<br />
+		- [Framework/Library 2]: [Configuration and usage context]<br />
+		- [Architectural Pattern 3]: [Implementation approach and reasoning]<br />
+		- [Environment Detail 4]: [Setup specifics and constraints]<br />
+
+		3. Codebase Status:<br />
+		- [File Name 1]:<br />
+		- Purpose: [Why this file is important to the project]<br />
+		- Current State: [Summary of recent changes or modifications]<br />
+		- Key Code Segments: [Important functions/classes with brief explanations]<br />
+		- Dependencies: [How this relates to other components]<br />
+		- [File Name 2]:<br />
+		- Purpose: [Role in the project]<br />
+		- Current State: [Modification status]<br />
+		- Key Code Segments: [Critical code blocks]<br />
+		- [Additional files as needed]<br />
+
+		4. Problem Resolution:<br />
+		- Issues Encountered: [Technical problems, bugs, or challenges faced]<br />
+		- Solutions Implemented: [How problems were resolved and reasoning]<br />
+		- Debugging Context: [Ongoing troubleshooting efforts or known issues]<br />
+		- Lessons Learned: [Important insights or patterns discovered]<br />
+
+		5. Progress Tracking:<br />
+		- Completed Tasks: [What has been successfully implemented with status indicators]<br />
+		- Partially Complete Work: [Tasks in progress with current completion status]<br />
+		- Validated Outcomes: [Features or code confirmed working through testing]<br />
+
+		6. Active Work State:<br />
+		- Current Focus: [Precisely what was being worked on in most recent messages]<br />
+		- Recent Context: [Detailed description of last few conversation exchanges]<br />
+		- Working Code: [Code snippets being modified or discussed recently]<br />
+		- Immediate Context: [Specific problem or feature being addressed before summary]<br />
+
+		7. Recent Operations:<br />
+		- Last Agent Commands: [Specific tools/actions executed just before summarization with exact command names]<br />
+		- Tool Results Summary: [Key outcomes from recent tool executions - truncate long results but keep essential info]<br />
+		- Pre-Summary State: [What the agent was actively doing when token budget was exceeded]<br />
+		- Operation Context: [Why these specific commands were executed and their relationship to user goals]<br />
+
+		8. Continuation Plan:<br />
+		- [Pending Task 1]: [Details and specific next steps with verbatim quotes]<br />
+		- [Pending Task 2]: [Requirements and continuation context]<br />
+		- [Priority Information]: [Which tasks are most urgent or logically sequential]<br />
+		- [Next Action]: [Immediate next step with direct quotes from recent messages]<br />
+	</Tag><br />
+
+	## Quality Guidelines<br />
+
+	- **Precision**: Include exact filenames, function names, variable names, and technical terms<br />
+	- **Completeness**: Capture all context needed to continue without re-reading the full conversation<br />
+	- **Clarity**: Write for someone who needs to pick up exactly where the conversation left off<br />
+	- **Verbatim Accuracy**: Use direct quotes for task specifications and recent work context<br />
+	- **Technical Depth**: Include enough detail for complex technical decisions and code patterns<br />
+	- **Logical Flow**: Present information in a way that builds understanding progressively<br />
+
+	This summary should serve as a comprehensive handoff document that enables seamless continuation of all active work streams while preserving the full technical and contextual richness of the original conversation.<br />
+</>;
 
 /**
  * Prompt used to summarize conversation history when the context window is exceeded.
@@ -50,106 +156,7 @@ export class ConversationHistorySummarizationPrompt extends PromptElement<Conver
 		return (
 			<>
 				<SystemMessage priority={this.props.priority}>
-					Your task is to create a comprehensive, detailed summary of the entire conversation that captures all essential information needed to seamlessly continue the work without any loss of context. This summary will be used to compact the conversation while preserving critical technical details, decisions, and progress.<br />
-
-					## Recent Context Analysis<br />
-
-					Pay special attention to the most recent agent commands and tool executions that led to this summarization being triggered. Include:<br />
-					- **Last Agent Commands**: What specific actions/tools were just executed<br />
-					- **Tool Results**: Key outcomes from recent tool calls (truncate if very long, but preserve essential information)<br />
-					- **Immediate State**: What was the system doing right before summarization<br />
-					- **Triggering Context**: What caused the token budget to be exceeded<br />
-
-					## Analysis Process<br />
-
-					Before providing your final summary, wrap your analysis in `&lt;analysis&gt;` tags to organize your thoughts systematically:<br />
-
-					1. **Chronological Review**: Go through the conversation chronologically, identifying key phases and transitions<br />
-					2. **Intent Mapping**: Extract all explicit and implicit user requests, goals, and expectations<br />
-					3. **Technical Inventory**: Catalog all technical concepts, tools, frameworks, and architectural decisions<br />
-					4. **Code Archaeology**: Document all files, functions, and code patterns that were discussed or modified<br />
-					5. **Progress Assessment**: Evaluate what has been completed vs. what remains pending<br />
-					6. **Context Validation**: Ensure all critical information for continuation is captured<br />
-					7. **Recent Commands Analysis**: Document the specific agent commands and tool results from the most recent operations<br />
-
-					## Summary Structure<br />
-
-					Your summary must include these sections in order, following the exact format below:<br />
-
-					<Tag name='analysis'>
-						[Chronological Review: Walk through conversation phases: initial request → exploration → implementation → debugging → current state]<br />
-						[Intent Mapping: List each explicit user request with message context]<br />
-						[Technical Inventory: Catalog all technologies, patterns, and decisions mentioned]<br />
-						[Code Archaeology: Document every file, function, and code change discussed]<br />
-						[Progress Assessment: What's done vs. pending with specific status]<br />
-						[Context Validation: Verify all continuation context is captured]<br />
-						[Recent Commands Analysis: Last agent commands executed, tool results (truncated if long), immediate pre-summarization state]<br />
-					</Tag><br />
-
-					<Tag name='summary'>
-						1. Conversation Overview:<br />
-						- Primary Objectives: [All explicit user requests and overarching goals with exact quotes]<br />
-						- Session Context: [High-level narrative of conversation flow and key phases]<br />
-						- User Intent Evolution: [How user's needs or direction changed throughout conversation]<br />
-
-						2. Technical Foundation:<br />
-						- [Core Technology 1]: [Version/details and purpose]<br />
-						- [Framework/Library 2]: [Configuration and usage context]<br />
-						- [Architectural Pattern 3]: [Implementation approach and reasoning]<br />
-						- [Environment Detail 4]: [Setup specifics and constraints]<br />
-
-						3. Codebase Status:<br />
-						- [File Name 1]:<br />
-						- Purpose: [Why this file is important to the project]<br />
-						- Current State: [Summary of recent changes or modifications]<br />
-						- Key Code Segments: [Important functions/classes with brief explanations]<br />
-						- Dependencies: [How this relates to other components]<br />
-						- [File Name 2]:<br />
-						- Purpose: [Role in the project]<br />
-						- Current State: [Modification status]<br />
-						- Key Code Segments: [Critical code blocks]<br />
-						- [Additional files as needed]<br />
-
-						4. Problem Resolution:<br />
-						- Issues Encountered: [Technical problems, bugs, or challenges faced]<br />
-						- Solutions Implemented: [How problems were resolved and reasoning]<br />
-						- Debugging Context: [Ongoing troubleshooting efforts or known issues]<br />
-						- Lessons Learned: [Important insights or patterns discovered]<br />
-
-						5. Progress Tracking:<br />
-						- Completed Tasks: [What has been successfully implemented with status indicators]<br />
-						- Partially Complete Work: [Tasks in progress with current completion status]<br />
-						- Validated Outcomes: [Features or code confirmed working through testing]<br />
-
-						6. Active Work State:<br />
-						- Current Focus: [Precisely what was being worked on in most recent messages]<br />
-						- Recent Context: [Detailed description of last few conversation exchanges]<br />
-						- Working Code: [Code snippets being modified or discussed recently]<br />
-						- Immediate Context: [Specific problem or feature being addressed before summary]<br />
-
-						7. Recent Operations:<br />
-						- Last Agent Commands: [Specific tools/actions executed just before summarization with exact command names]<br />
-						- Tool Results Summary: [Key outcomes from recent tool executions - truncate long results but keep essential info]<br />
-						- Pre-Summary State: [What the agent was actively doing when token budget was exceeded]<br />
-						- Operation Context: [Why these specific commands were executed and their relationship to user goals]<br />
-
-						8. Continuation Plan:<br />
-						- [Pending Task 1]: [Details and specific next steps with verbatim quotes]<br />
-						- [Pending Task 2]: [Requirements and continuation context]<br />
-						- [Priority Information]: [Which tasks are most urgent or logically sequential]<br />
-						- [Next Action]: [Immediate next step with direct quotes from recent messages]<br />
-					</Tag><br />
-
-					## Quality Guidelines<br />
-
-					- **Precision**: Include exact filenames, function names, variable names, and technical terms<br />
-					- **Completeness**: Capture all context needed to continue without re-reading the full conversation<br />
-					- **Clarity**: Write for someone who needs to pick up exactly where the conversation left off<br />
-					- **Verbatim Accuracy**: Use direct quotes for task specifications and recent work context<br />
-					- **Technical Depth**: Include enough detail for complex technical decisions and code patterns<br />
-					- **Logical Flow**: Present information in a way that builds understanding progressively<br />
-
-					This summary should serve as a comprehensive handoff document that enables seamless continuation of all active work streams while preserving the full technical and contextual richness of the original conversation.<br />
+					{SummaryPrompt}
 				</SystemMessage>
 				{history}
 				{this.props.workingNotebook && <WorkingNotebookSummary priority={this.props.priority - 2} notebook={this.props.workingNotebook} />}
@@ -383,6 +390,8 @@ class ConversationHistorySummarizer {
 		@ILogService private readonly logService: ILogService,
 		@IInstantiationService private readonly instantiationService: IInstantiationService,
 		@IConfigurationService private readonly configurationService: IConfigurationService,
+		@IExperimentationService private readonly experimentationService: IExperimentationService,
+		@IEndpointProvider private readonly endpointProvider: IEndpointProvider,
 	) { }
 
 	async summarizeHistory(): Promise<{ summary: string; toolCallRoundId: string }> {
@@ -427,11 +436,24 @@ class ConversationHistorySummarizer {
 
 	private async getSummary(mode: SummaryMode, propsInfo: ISummarizedConversationHistoryInfo): Promise<FetchSuccess<string>> {
 		const stopwatch = new StopWatch(false);
-		const endpoint = this.props.endpoint;
+		const forceGpt41 = this.configurationService.getExperimentBasedConfig(ConfigKey.Internal.AgentHistorySummarizationForceGpt41, this.experimentationService);
+		const endpoint = forceGpt41 ?
+			await this.endpointProvider.getChatEndpoint('gpt-4.1') :
+			this.props.endpoint;
 
 		let summarizationPrompt: ChatMessage[];
+		const promptCacheMode = this.configurationService.getExperimentBasedConfig(ConfigKey.Internal.AgentHistorySummarizationWithPromptCache, this.experimentationService);
 		try {
-			summarizationPrompt = (await renderPromptElement(this.instantiationService, endpoint, ConversationHistorySummarizationPrompt, { ...propsInfo.props, simpleMode: mode === SummaryMode.Simple }, undefined, this.token)).messages;
+			if (mode === SummaryMode.Full && promptCacheMode) {
+				const props: AgentPromptProps = {
+					...propsInfo.props,
+					triggerSummarize: false
+				};
+				const expandedEndpoint = endpoint.cloneWithTokenOverride(endpoint.modelMaxPromptTokens * 1.05);
+				summarizationPrompt = (await renderPromptElement(this.instantiationService, expandedEndpoint, AgentPromptWithSummaryPrompt, props, undefined, this.token)).messages;
+			} else {
+				summarizationPrompt = (await renderPromptElement(this.instantiationService, endpoint, ConversationHistorySummarizationPrompt, { ...propsInfo.props, simpleMode: mode === SummaryMode.Simple }, undefined, this.token)).messages;
+			}
 			this.logInfo(`summarization prompt rendered in ${stopwatch.elapsed()}ms.`, mode);
 		} catch (e) {
 			const budgetExceeded = e instanceof BudgetExceededError;
@@ -461,7 +483,12 @@ class ConversationHistorySummarizer {
 				),
 			} : undefined;
 
-			stripCacheBreakpoints(summarizationPrompt);
+			if (promptCacheMode) {
+				addCacheBreakpoints(summarizationPrompt);
+			} else {
+				stripCacheBreakpoints(summarizationPrompt);
+			}
+
 			summaryResponse = await endpoint.makeChatRequest(`summarizeConversationHistory-${mode}`, ToolCallingLoop.stripInternalToolCallIds(summarizationPrompt), undefined, this.token ?? CancellationToken.None, ChatLocation.Other, undefined, {
 				temperature: 0,
 				stream: false,
@@ -584,6 +611,17 @@ class ConversationHistorySummarizer {
 			promptCacheTokenCount: usage?.prompt_tokens_details.cached_tokens,
 			responseTokenCount: usage?.completion_tokens,
 		});
+	}
+}
+
+export class AgentPromptWithSummaryPrompt extends PromptElement<AgentPromptProps> {
+	override async render(state: void, sizing: PromptSizing) {
+		return <>
+			<AgentPrompt {...this.props} />
+			<UserMessage>
+				{SummaryPrompt}
+			</UserMessage>
+		</>;
 	}
 }
